add_files not correctly matching file extension
The add_files function does not limit itself to the given fileExtension but matches any occurrence of that string at the end of a filename. For example, example.fileName and exampleFile.Name would both be matched if the fileExtenion was set to "Name". The intended functionality should only match exampleFile.Name.