Added multiple="all" to a full_join() that caused warnings in many tests
This commit message is to reassure myself that I am not breaking
anything. After all, failing unit tests are supposed to mean that
something has been broken. But that is not the case here. This fix is
only to suppress warnings, everything has already been working as
intended.

dplyr::full_join() threw a warning that:

Each row in `x` is expected to match at most 1 row in `y`.

I suppose this warning has been added to dplyr after the last release of
emuR (2.3.0). It can be silenced by adding multiple="all" to the
function call if multiple matches are expected.

After analysing the function bundleAnnotDFsToAnnotJSONchar, I am now
sure that the behavior is expected: x contains annotation items, and y
contains level/attribute pairs, and x and y are matched by level. As per
the design of EMU-SDMS, levels are allowed to have one or more
attributes; so we have to expect multiple matches for every item in x
whose level has more than one attribute.