follow-up to 8d342079: adapt to tidyverse 2.0.0
The previous commit was prepared with Tidyverse 1.3.2, although 2.0.0 is
the current version. We now go with the many-to-many hint instead of
multiple="all".

Hint: We must expect multiple matches in both directions. Multiple y
matches for each row in x, because levels can have multiple attributes.
And multiple x matches for each row in y, because we can have multiple
items per level in one go.