readr::write_lines is multi-threaded by default and can cause problems
We started seeing random-looking test failures on CRAN on August 17. We
have now found out that on our Windows machine, write_lines with
multiple threads and cause R to hang forever. This happens only with
some input data, and even with the same data only in one of about 300
cases.

This would also be fixed with the very recent change to readr/vroom:
https://github.com/tidyverse/vroom/pull/512

But that fix is not on CRAN yet.