Users can now provide end times of SEGMENTS to create_itemsInLevel()
The function create_itemsInLevel started out only supporting levels of
type ITEM. Later, it gained support for levels of type EVENT and
SEGMENT.

However, SEGMENT levels were only supported in a particular way: Users
could only provide the start time of each segment, but not the end time.
The end time was filled in based on the start time of the next neighbor.
The end time of the last segment was filled in based on the media file
length.

Now, we add a new option. Users can explicitly provide the end time for
each segment.

This is necessary if users want to use a segment list obtained via
query(), modify it and feed it to create_itemsInLevel(). If they did
that before, the end time of those segment lists  was silently ignored
and replaced.